refactor(cli): extract gate receipt preparation
